# src/api-docs.json
{"openapi":"3.0.1","servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/room/{id}":{"get":{"tags":["room-controller"],"operationId":"getCustomerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDTO"}}}}}},"put":{"tags":["room-controller"],"operationId":"updateRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDTO"}}}}}},"delete":{"tags":["room-controller"],"operationId":"deleteRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/reservation/{id}":{"put":{"tags":["reservation-controller"],"description":"Update reservation info service","operationId":"updateReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}},"required":true},"responses":{"202":{"description":"Successfully Update reservation info!","content":{"application/json":{"example":{"code":202,"Status":"Accepted!","Message":"Successfully Update reservation info!"}}}}},"security":[{"token":[]}]},"delete":{"tags":["reservation-controller"],"description":"Delete Reservation service","operationId":"deleteReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully Reservation customer!","content":{"application/json":{"example":{"code":200,"Status":"Accepted!","Message":"Successfully delete Reservation!"}}}}},"security":[{"token":[]}]}},"/api/v1/employee/{id}":{"get":{"tags":["employee-controller"],"description":"Get Employee by id service","operationId":"getEmployeeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully get Employee by id!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get Employee by id!"}}}}},"security":[{"token":[]}]},"put":{"tags":["employee-controller"],"description":"Update Employee info service","operationId":"updateEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}},"required":true},"responses":{"202":{"description":"Successfully Update Employee info!","content":{"application/json":{"example":{"code":202,"Status":"Accepted!","Message":"Successfully Update Employee info!"}}}}},"security":[{"token":[]}]},"delete":{"tags":["employee-controller"],"description":"Delete Employee service","operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully delete Employee!","content":{"application/json":{"example":{"code":200,"Status":"Accepted!","Message":"Successfully delete Employee!"}}}}},"security":[{"token":[]}]}},"/api/v1/customers/{id}":{"get":{"tags":["customer-controller"],"description":"Get customer by id service","operationId":"getCustomerById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully get customer by id!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get customer by id!"}}}}},"security":[{"token":[]}]},"put":{"tags":["customer-controller"],"description":"Update customer info service","operationId":"updateCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}},"required":true},"responses":{"202":{"description":"Successfully Update customer info!","content":{"application/json":{"example":{"code":202,"Status":"Accepted!","Message":"Successfully Update customer info!"}}}}},"security":[{"token":[]}]},"delete":{"tags":["customer-controller"],"description":"Delete customer service","operationId":"deleteCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully delete customer!","content":{"application/json":{"example":{"code":200,"Status":"Accepted!","Message":"Successfully delete customer!"}}}}},"security":[{"token":[]}]}},"/api/v1/billing/{id}":{"get":{"tags":["billing-controller"],"description":"Get Billing by id service","operationId":"getBillingById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully get Billing by id!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get Billing by id!"}}}}},"security":[{"token":[]}]},"put":{"tags":["billing-controller"],"description":"Update Billing info service","operationId":"updateBilling","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillingDTO"}}},"required":true},"responses":{"202":{"description":"Successfully Update Billing info!","content":{"application/json":{"example":{"code":202,"Status":"Accepted!","Message":"Successfully Update Billing info!"}}}}},"security":[{"token":[]}]}},"/api/v1/HousekeepingTasks/{id}":{"get":{"tags":["housekeeping-tasks-controller"],"description":"Get HousekeepingTasks by id service","operationId":"HousekeepingTasksById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully get HousekeepingTasks by id!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get HousekeepingTasks by id!"}}}}},"security":[{"token":[]}]},"put":{"tags":["housekeeping-tasks-controller"],"description":"Update HousekeepingTasks info service","operationId":"updateHousekeepingTasksDTO","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HousekeepingTasksDTO"}}},"required":true},"responses":{"202":{"description":"Successfully Update HousekeepingTasks info!","content":{"application/json":{"example":{"code":202,"Status":"Accepted!","Message":"Successfully Update HousekeepingTasks info!"}}}}},"security":[{"token":[]}]},"delete":{"tags":["housekeeping-tasks-controller"],"description":"Delete HousekeepingTasks service","operationId":"deleteHousekeepingTasks","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully delete HousekeepingTasks!","content":{"application/json":{"example":{"code":200,"Status":"Accepted!","Message":"Successfully delete HousekeepingTasks!"}}}}},"security":[{"token":[]}]}},"/api/v1/room":{"get":{"tags":["room-controller"],"operationId":"getAllRoom","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomDTO"}}}}}}},"post":{"tags":["room-controller"],"operationId":"createRoom","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDTO"}}}}}}},"/api/v1/reservation":{"get":{"tags":["reservation-controller"],"description":"Get all Reservation service","operationId":"getAllReservation","responses":{"200":{"description":"Successfully get all Reservation!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get all Reservation!"}}}}},"security":[{"token":[]}]},"post":{"tags":["reservation-controller"],"description":"Create reservation service","operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}},"required":true},"responses":{"201":{"description":"Successfully create reservation!","content":{"application/json":{"example":{"code":201,"Status":"Created!","Message":"Successfully create reservation!"}}}}},"security":[{"token":[]}]}},"/api/v1/employee":{"get":{"tags":["employee-controller"],"description":"Get all Employee service","operationId":"getAllEmployee","responses":{"200":{"description":"Successfully get all Employee!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get all Employee!"}}}}},"security":[{"token":[]}]},"post":{"tags":["employee-controller"],"description":"Create Employee service","operationId":"createEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}},"required":true},"responses":{"201":{"description":"Successfully create Employee!","content":{"application/json":{"example":{"code":201,"Status":"Created!","Message":"Successfully create Employee!"}}}}},"security":[{"token":[]}]}},"/api/v1/customers":{"get":{"tags":["customer-controller"],"description":"Get all customers service","operationId":"getAllCustomers","responses":{"200":{"description":"Successfully get all customers!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get all customers!"}}}}},"security":[{"token":[]}]},"post":{"tags":["customer-controller"],"description":"Create customer service","operationId":"createCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}},"required":true},"responses":{"201":{"description":"Successfully create customers!","content":{"application/json":{"example":{"code":201,"Status":"Created!","Message":"Successfully create customer!"}}}}},"security":[{"token":[]}]}},"/api/v1/billing":{"get":{"tags":["billing-controller"],"description":"Get all Billing service","operationId":"getAllBilling","responses":{"200":{"description":"Successfully get all Billing!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get all Billing!"}}}}},"security":[{"token":[]}]},"post":{"tags":["billing-controller"],"description":"Create Billing service","operationId":"createBilling","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillingDTO"}}},"required":true},"responses":{"201":{"description":"Successfully create Billing!","content":{"application/json":{"example":{"code":201,"Status":"Created!","Message":"Successfully create Billing!"}}}}},"security":[{"token":[]}]}},"/api/v1/auth/register":{"post":{"tags":["authentication-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/auth/refresh-token":{"post":{"tags":["authentication-controller"],"operationId":"refreshToken","responses":{"200":{"description":"OK"}}}},"/api/v1/auth/authenticate":{"post":{"tags":["authentication-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/HousekeepingTasks":{"get":{"tags":["housekeeping-tasks-controller"],"description":"Get all HousekeepingTasks service","operationId":"getAllHousekeepingTasks","responses":{"200":{"description":"Successfully get all HousekeepingTasks!","content":{"application/json":{"example":{"code":200,"Status":"Ok!","Message":"Successfully get all HousekeepingTasks!"}}}}},"security":[{"token":[]}]},"post":{"tags":["housekeeping-tasks-controller"],"description":"Create HousekeepingTasks service","operationId":"createHousekeepingTasksDTO","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HousekeepingTasksDTO"}}},"required":true},"responses":{"201":{"description":"Successfully create HousekeepingTasks!","content":{"application/json":{"example":{"code":201,"Status":"Created!","Message":"Successfully create HousekeepingTasks!"}}}}},"security":[{"token":[]}]}},"/api/v2/room":{"get":{"tags":["room-controller-v-2"],"operationId":"getAllRooms","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomDTO"}}}}}}}},"/api/v1/room/customer/{id}":{"get":{"tags":["room-controller"],"operationId":"getCustomerById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/room/available":{"get":{"tags":["room-controller"],"operationId":"findAvailableRoom","parameters":[{"name":"available","in":"query","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomDTO"}}}}}}}},"/api/v1/reservation/name":{"get":{"tags":["reservation-controller"],"description":"find Reservation By Customer Name","operationId":"getReservationByCustomerName","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Successfully find Reservation By Customer Name","content":{"application/json":{"example":{"code":201,"Status":"Created!","Message":"Successfully find Reservation By Customer Name"}}}}},"security":[{"token":[]}]}},"/api/v1/reservation/CustomerID":{"get":{"tags":["reservation-controller"],"description":"find Reservation By Customer id","operationId":"getReservationByCustomerID","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"201":{"description":"Successfully find Reservation By Customer id","content":{"application/json":{"example":{"code":201,"Status":"Created!","Message":"Successfully find Reservation By Customer id"}}}}},"security":[{"token":[]}]}},"/api/v1/customers/name/{name}":{"get":{"tags":["customer-controller"],"description":"find customer by name","operationId":"findCustomerByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Successfully find customer by name","content":{"application/json":{"example":{"code":201,"Status":"Created!","Message":"Successfully find customer by name"}}}}},"security":[{"token":[]}]}},"/api/rooms":{"get":{"tags":["room-controller-v-2"],"operationId":"getAllRoomsByHeader","parameters":[{"name":"API-Version","in":"header","required":false,"schema":{"type":"string","default":"1"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomDTO"}}}}}}}},"/api/room":{"get":{"tags":["room-controller-v-2"],"operationId":"getAllRoomsByQuery","parameters":[{"name":"version","in":"query","required":false,"schema":{"type":"string","default":"1"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomDTO"}}}}}}}}},"components":{"schemas":{"RoomDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"available":{"type":"boolean"},"status":{"type":"string"},"price":{"type":"number","format":"double"},"facilities":{"type":"string"},"capacity":{"type":"integer","format":"int32"},"size":{"type":"number","format":"double"}}},"ReservationDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"checkIn":{"type":"string","format":"date-time"},"checkOut":{"type":"string","format":"date-time"},"numberOfCheckers":{"type":"integer","format":"int32"},"customerId":{"type":"integer","format":"int64"},"customerName":{"type":"string"},"roomId":{"type":"integer","format":"int64"},"reservationDate":{"type":"string","format":"date-time"}}},"EmployeeDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"CustomerDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"BillingDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"invoiceDate":{"type":"string","format":"date-time"},"totalAmount":{"type":"integer","format":"int32"},"status":{"type":"string"},"reservationId":{"type":"integer","format":"int64"}}},"HousekeepingTasksDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"taskDescription":{"type":"string"},"taskStatus":{"type":"string"},"scheduledDate":{"type":"string","format":"date-time"},"completedDate":{"type":"string","format":"date-time"},"roomId":{"type":"integer","format":"int64"},"employeeId":{"type":"integer","format":"int64"}}},"RegisterRequest":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"}}},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}